cadetapp/requirements.txt
Django==1.11.6
pytz==2017.2
django-appconf==1.0.2
django-bower==5.2.0
django-chartjs==1.2
django-compressor==2.2
django-libsass==0.7
django-nvd3==0.9.7
django-pipeline==1.6.13
django-sass-processor==0.5.5
django-storages==1.6.5
matplotlib==2.0.2
numpy==1.13.3
numpydoc==0.7.0
panda==0.3.1
pandas==0.20.3
djangorestframework==3.7.3
pipenv==8.3.2
requests==2.9.1
<<<<<<< HEAD
polling==0.3.0
jsonfield==2.0.2
=======

coverage==3.6
selenium==2.33.0
>>>>>>> 373445e9d96cc258a4b85e94a929907d22b2a32e
